Add files via upload
it's day 5 today i have completed half_adder in which two input's will be add and it produce sum and carry where when we add some different number like a =0 ,b=1 then sum will 0 and carry will 1 and when we add same number as a=1 b=1 then sum=0, carry = 1. advantages of half adder  is it can be use in alu(arithmetic logical unit), logic operators so on. 

but there is also some disadvantages like it can hold  carry inputs  from previous stages which limits they use single bits .